+++ modules/core/src/main/scala/turbolift/interpreter/Features_opaque.scala
@@ -3,5 +3,5 @@ package turbolift.interpreter
 
 private[interpreter] trait Features_opaque:
   extension (thiz: Features)
-    inline def |(that: Features): Features = Features.wrap(thiz.unwrap | that.unwrap)
-    inline def mask: Features = Features.wrap(thiz.unwrap & Features.Mask.unwrap)
+    inline def |(that: Features): Features = Features.wrap((thiz.unwrap | that.unwrap).toByte)
+    inline def mask: Features = Features.wrap((thiz.unwrap & Features.Mask.unwrap).toByte)
